requirements.txt
django
djangorestframework
psycopg2-binary
python-decouple
pillow
django-filter
djangorestframework-simplejwt
dj-config-url
drf-yasg